Add files via upload
Play a song from this repl and build a menu to go with it. Make it look like an iPod!

When you run the code below, it should open the Output tool which will have a checkbox with a headphones icon in the bottom right corner. Select this checkbox to hear the sound. (It may be behind the video.)