Add files via upload
Build a really cool to do list manager. (I know we did this before...hold on!)

We are going to upgrade the last to do list manager we created:

Create a menu where the user can view, add, or remove an item.
The user should be able to edit the text of an item on the list too.
Don't allow the user to add duplicates.
Double check with the user they want to remove an item from the list before it is actually removed. (Is this the item they really want to remove?)
Give the user the option to completely erase the to do list. (You should be able to do this in one line of code!)
Example:

To Do List Manager:
Do you want to view, add, edit, or remove an item from the to do list?
view
record the video for day 36
Do you want to view, add, edit, or remove an item from the to do list?
remove
What do you want to remove?
record the video for day 36
Are you sure you want to remove this?
yes