Add files via upload
Made it? Good! Let's get cracking.

Your system should:

Have a menu that asks if you want to add, view, move or edit a 'to do'.

If you choose 'add' then the system should:

Prompt you to input what the to do is, when it is due by and the priority (high, medium or low).
Add the 'to do' to the list.
'View' should give two options:

View all - shows all 'to dos' with a pretty print.
View priority - allows you to search for high, medium or low priority and only see matching tasks.
'Edit' allows you to change any of the information within one of the 'to dos'.

'Remove' lets you completely remove a 'to do' when it is 'to done'.