Add files via upload
In yesterday's challenge, you created a file called high.score and stored some high scores in it.

We've added a version of that file to this repl.

Your program should:

Read in the data from the high.score file.
Identify which of those users had the highest score. Automatically! (Not you doing it!)
Output the name and score of the winner.