Add files via upload
Day 50! Boy, you are smashing this 100 days! 🎊 To celebrate, here's a project for you.

Your idea storage system should:

Prompt the user to add an idea, or load a random one.
Choosing 'Add an idea' results in:
A prompt for the user to input their idea.
Add the idea to a text file called 'my.ideas'.
Add the idea in append mode, with every new idea on a brand new line.
Choosing 'Load random' results in:
Load the list of ideas.
Choose one at random.
Display it on screen for a few seconds.
Return to the menu.