Add files via upload
There's no place like Rome...Or Napoli, Milan, possibly even New York if you must.

Just not the dodgy 2am 'round bread with suspicious toppings' merchants that I definitely do not visit on my way home from a night out.

That's right, you're opening a pizza shop! Try not to get anchovy on your keyboard. Man that stuff never washes out.

Regardless, your program today must:

Prompt the user to input the quantity and size of pizzas.

Multiply the two inputs together to calculate the cost of the pizzas.

Store that in a 2D list with the user's name.

Use try.... except for two reasons:

Include auto-save and auto-load. Use it with the auto-load.
When casting the quantity of pizzas to an integer. 
Avoid the user crashing the program by typing 'three'
 instead of '3'. Or any other non-integer input. If they do, then prompt them to try again.