Add files via upload
Your video game inventory system should:

Have a menu that allows the user to:
Add
View
Remove
Adding an item saves it to a file using captalize mode. Duplicates are allowed.
Removing an item deletes it from the file.
View is trickier. It should output the name of the item and tell you how many of those items you have.
Use auto-save and auto-load with try... except.