Add files via upload
You're going to write a program that checks a string to see if it is a palindrome.

Yes. We know that Python has the built in function string.reverse() that you could use.

Zero points for that today, we want you to think hard and utilize your skills in:

recursion
string slicing
looping
Your program should:

Prompt the user to input a word.
Analyze the word to see if it is a palindrome.
Output a relevant 'yes/no message.